Working reverse-window compression cache

+++ compression.h
@@ -32,7 +32,7 @@ class decompress {
     uint32_t compressedSize;
     uint32_t compressedPos;
     uint8_t* outCache;
-    uint16_t cacheLen;
+    uint32_t cacheLen;
     uint32_t cacheStart;
 #ifdef ENABLE_OEPLFS
     OEPLFile *file = nullptr;
